Fix bug in KernelPatchArg init in VLL
The KernelPatchArg structure was added to increase performance -
it removes some pointer lookups and conditionals in the NCC code.

VLL creates two args, one for the "normal" case and one for the
combined case. However, the combined case is only used if
"check_combined_WNCC" is set to True.

When doing matchtag regeneration, check_combined_WNCC will be
false, since matchtag regen is only ran at level 0 and
check_combined_WNCC is set to true only when the Pyramid_step
(i.e. level) is greater than Py_combined_level.
Py_combined_level is always zero, so check_combined_WNCC is always
false when doing matchtag regen.

So, the KernelPatchArg for combined WNCC in VLL, called patch_next,
is not used during matchtag regen. However, it is initialized.
The initialization dereferences the "Py_Images_next" and
"Py_MagImages_next" pointers. But these are NULL (they are only
set when (level > Py_combined_level)). So - we get a null pointer
dereference here.

This didn't happen during DEM generation because the levelinfo
structure that holds the "Py_Image_next" pointers had stale but
accessible pointer values. This is still a bug, but didn't lead
to a crash. It also didn't have any effect on the results since
"patch_next" is not used in level 0, which is the only level that
has the stale pointers.