Make pyramid image generation deterministic
A omp loop was added to the gaussian kernel generation that
included a sum reduction on a float. This resulted in non
deterministic behavior.

Removing the omp sections fixed the problem, but also slowed
down the code. This seemed a little strange -- the filter
size for the kernel is set to 9 or 15, and the function is called
on an order of a couple hundred times. It's a double nested
loop across the filter size, so the loop should only iterate
a couple hundred times.

Given the small amount of work that the loop does, it feels like
omp shouldn't result in a huge performance change.

Moving the gausian kernel generation to a new function seems to
alleviate the performance issue - resulting in about the same
performance on a small example as the non-deterministic omp
version. I have no idea why that's the case, but it seems to work
for cray.

Testing:
 - tested bench1-6 8m before and after this change. Saw
   non-deterministic results before change, results were
   deterministic after change. Tests were run on BW cray.