Add MPI work queue for tiles
Implement using MPI counter. Note: Need async progress env
defines in runscript for this to work correctly for some MPI
implementations.

Ensure TileIndexer subclass destructors are called

Need to have a virtual destructor on the TileIndexer to call
the destructors of it's subclasses.

MPI breaks if the MPITileIndexer does not have it's destructor
called, so this is pretty important