Make SetHeightRange_blunder deterministic
Previosly, SetHeightRange_blunder was nondeterminisitc.
In an openmp loop, it iterated over the triangles and interpolated
height values based on where grid points fell within individual
triangles. Some points fell in multiple triangles (on edges), so
height value is interpolated from multiple triangles. Different
trianlges yield different heights (though close) due to floating
point rounding.

To fix this, add an atomic height array that's used to store the
height values as we iterate through the triangles. If point is
in multiple triangles, use the max height value. This is updated
using a compare and swap.

Finally, copy the height values into the output array. We can't
just use the output array because it is not atomic, and further,
we can't track whether a value has already been updated based on
the initial value in the array, as there are no gaurentees about
the initial value of the array.