Fix bug in logging
MPI logging was broken after moving logging to a separate func.
The BUILDMPI compiler define didn't make it to the compilation
stage, so it was build without MPI support regardless of whether
or not it was an MPI build.

Fix this by refactoring the Makefile a bit. Note we could go
further here and move the setsm.o and setsm_mpi.o objects to
the respective object lists instead of manually specifying their
build commands.

Also, explicitly require initialization of logging. init_logging
must be called after mpi is initialized. doing it this way will
make errors more noisy.