Fix buffer overflow in Readtiff_T
There's a buffer overflow in readtiff_t. It occurs when
accessing the output array in the tile path of the function.

The output array is indexed using the following variables:
    row, col, tileL, i, j, data_size->width

These are all either ints or unsigned ints. On Bluewaters,
they all are 32 bits in length.

This poses a problem when trying to read images whose
width*height is greater than somewhere around 2^31 or 2^32.

In turns out this case is hit fairly often, leading to an
overflow and often a crash.

Fix this by changing i and j to size_t, which should be
large enough to index any array. I'm not positive that this
is good enough for **all** cases, but it might be.

I didn't just change all variables to size_t because there
are size requirements in the signatures for libtiff for
the tileL and the tileW variables, and also the "row" variable,
since it's used in the scanline code.