Add check for async progress support
This hopefully prevents the case where async progress support
is not available, but setsm is run in mpi-mode anyway. This
would be pretty inefficient, as the other ranks would not
be able to dequeue a tile until rank 0 got around to interacting
with the queue.

This change made the mpi-init code pretty long, so I moved it
to the mpi_helpers header in an attempt to not clutter up main
too much.

Additioanlly, since we previously moved the MPI initialization
to main (it was previously elsewhere...) we need to make sure
we finalize MPI in main as well.