Fix build error on gcc/4.8.0
There was some preprocessor logic in setsm from the C days
that ensured the bool type was defined.

This was in Typedefine.hpp, and took the form of a preprocessor
"ifndef". At some point, some other defines were added to that
"ifndef" that were not related to the bool type.

Our compilers don't define a "bool" macro at the preprocessor
level, so these defines are always included.

It appears that depending on header order for gcc 4.8.0, and
other gcc versions, a bool macro may be defined.

When that happens, these other macros, like pwrtwo and SWAP,
are not defined. This leads to build errors.

Since we now target c++11, we can just remove this logic.
"bool" is a keyword, so it'll always be there as a type.

This commit also fixes some type conversion warnings
that come up from no longer defining our own true and false.
The logic should hopefully still be the same.