Move pair check in vll to right before INCC update
If check_matching_rate is set to true, then the GridVoxel
is not initialized, so we cannot do the early pair check.

We don't want to do the check_mathing_rate where the pair
check was, because we want to update the peaks if it's set
to true, so we don't want to skip it.

So, just move the check to later in the processing.