Only convert NODATA values to zero for input image
To implement this, we only do the conversion if the image data
type is uint16_t. DEMs use floats, and other output files use
uint8_t. So far, the image sources we've used use the type
uint16_t, so this should be fine for now.

Geotiff files have special pixel values to represent "there's no
valid data here". This value can vary, and is stored in a TIFF
TAG called GDAL_NODATA.

SETSM would like these pixels with no data in them to have values
of zero. SO, if a non-zero value is used to indicate that a pixel
has no data, we convert it to zero.

However, we don't want to do this with DEMs. DEMs use -9999 for
the NODATA value. Zero is a valid DEM height, so we cannot use
it as the NODATA value, nor do we want to convert -9999 to zeros
since they would then show up as valid points.