Must define static variable in compilation unit
Without this, we get an error on cray. This doesn't happen with
GNU and intel on OSC systems because they optimize it away, so
there's no need for the actual definition. C++17 has the inline
keyword that fixes this issue, but I think it's best to stick
to c++11 for now