Change data structure for Grid Pair
Grid pair tracks image pairs for each grid point. It stores
integer ID values that map to pairs, which are currently
stored in the PairInfo structure.

It was implemented as a vector of vectors. However, since
lots of points share the same pairs, this structure had alot
of redundant data in it. Additionally, there was lots of
overhead, since each point had dynamically allocated memory
associated with it (the inner vector).

This commit changes that implementation. There is one vector
over all grid points, which holds ids that map to pair lists.
The pair lists are stored in a map with no duplication.

In practice this seems to reduce memory usage, and greatly
reduces the number of dynamic memory allocations.

Conflicts:
	setsm_code.cpp
	setsm_code.hpp