Fix inconsistencies in selected reference image
select_referenceimage is used to find out what the reference
image is for a given grid point. It's called in actual_pair and
in vertical line locus. It's important for these to return
the same value for a given pair.

Why? Because of the hopefully coming soon changes to grid voxel
where we only allocate the space necessary for each pair. To
do this, we need to know which pair numbers are used at each
grid point, and knowing the reference image is part of this.

select_reference image takes a min and max height as arguments.
So, ensure that we use the same min and max heights at each call
site.

In particalar, MJ suggests using the ori_minmaxHeight values.