Fix consistency issue in coordinate calculation
in SetOrthoImageCoord, the coordinate calculation looked somewhat
like this:

f(image_resolution_grid_coordinate) -> dem_grid_coordinate

And the reverse was done in various VerticalLineLocus_blunder
functions:

g(dem_grid_coordinate) -> image_resolution_grid_coordiante

Where g is roughly the inverse of f.

However, this did not work very well in practice. There were
off-by-1 errors due to floating point rounding.

So, we do the same calculation in both places. It might make
sense to extract this to it's own function/module eventually,
but for now I think this works.

We keep the conversion in SetOrthoImageCoord more-or-less the
same. It starts with an image_resolution_grid_coordinate and
converts that to a dem_grid_coordiante.

But, we change the conversion in VLL-blunder. First we find
the image coordinate for the dem grid coordinate. Then, from
that image coordinate, we find an "aligned" dem grid coordinate,
and use this value.

In mathy-pseudocode, this now looks like:

im_coord := g(dem_grid_coordiante)
aligned_dem_grid_coordinate := f(im_coord)

This "im_coord" is used as the base for all calculations, so
that the math is the same as in SetOrthoImageCoords.

Also removed are some redundant values in the calculation (the
boundary values).