Use running avg in tile merge to save memory
Previously, a list of DEM values for each DEM grid point
was kept. This lead to a large memory spike at the end of
the program (sometimes > 100GB).

Instead of saving a list, compute the average as we go. This
requires only 16-bytes per DEM grid point. An empty vector will
use at least this much memory.