Free SubImages, SubMagImages, SubOriImages
Previouly the level variable was used to index into these
variables. However, the level variable was updated before
the "freeing" loop, so it did not iterate the correct
number of times.

The level variable is initialized from the Pyramid_step
variable, so just use that in the loops (it's const).